# backend/app/schemas/common.py
from datetime import datetime
from typing import Generic, TypeVar

from pydantic import BaseModel, ConfigDict, Field


class ORMModel(BaseModel):
    model_config = ConfigDict(from_attributes=True)


T = TypeVar("T")


class Page(BaseModel, Generic[T]):
    items: list[T]
    total: int
    page: int = Field(ge=1)
    page_size: int = Field(ge=1, le=200)


class Message(BaseModel):
    message: str


class HealthResponse(BaseModel):
    status: str
    database: str
    face_engine: str
    time: datetime